:recycle: Fix extractor
